+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package layered_hardware_gazebo
 0.0.8 (next version)
 --------------------
 * Support setting the initial position of joints
+* Add fixed_pid operation mode
 
 0.0.7 (2020-05-08)
 ------------------